Revert "Fixed an issue where image quality was not being set"
