+++ .gitignore
@@ -90,5 +90,3 @@ typings/
 lib/
 
 package-lock.json
-
-output.jpg
