+++ package.json
@@ -1,17 +1,12 @@
 {
   "name": "@shopify/screenshot-glb",
-  "version": "1.3.3",
+  "version": "1.3.2",
   "publishConfig": {
     "access": "public"
   },
   "description": "A command line utility for taking screenshots of a glTF 2.0 binary file using Google's <model-viewer> component.",
   "bin": {
     "screenshot-glb": "./src/cli.js"
-  },
-  "scripts": {
-    "lint": "echo \"linting not implemented\"",
-    "test": "jest",
-    "dev": "node src/cli.js -i ./test/fixtures/Astronaut.glb -o output.jpg -f image/jpeg -q 1.00 -v -t 30000"
   },
   "author": "Stephan Leroux <stephan.leroux@shopify.com>",
   "license": "MIT",
@@ -20,8 +15,5 @@
     "mime-types": "^2.1.27",
     "puppeteer": "^5.3.1",
     "yargs": "^13.3.0"
-  },
-  "devDependencies": {
-    "jest": "^26.6.3"
   }
 }
