+++ src/colors.js
@@ -1,4 +1,4 @@
 module.exports = {
-  white: 'rgba(255, 255, 255, 1)',
+  white: 'rgba(255, 255, 255, 100)',
   transparent: 'rgba(255, 255, 255, 0)',
 }
